% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guaguas.R
\docType{data}
\name{guaguas}
\alias{guaguas}
\title{Nombres de bebés (guaguas)}
\format{
Un dataframe con 84565 filas y 4 variables
\describe{
\item{anio}{Año}
\item{sexo}{Sexo registral}
\item{nombre}{Nombre}
\item{n}{Frecuencia del nombre ese año}
}
}
\source{
Servicio de Registro Civil e Identificación
de Chile
}
\usage{
guaguas
}
\description{
Set de datos con nombres de bebés registrados en Chile
entre 1920 y 2019, según el Servicio de Registro Civil
e Identificación. Incluye todos los nombres con al
menos 15 ocurrencias.
}
\keyword{datasets}
